COMP: Support configuring project specifying custom CMAKE_MACOSX_RPATH value
This commit updates the build-system to set CMAKE_MACOSX_RPATH to TRUE
only if is not already defined in the current scope. The default
behavior remains unchanged.

This allows to successfully configure the project with the following
options so that RPATH with absolute paths is used for both the build
and install trees:

  -DCMAKE_MACOSX_RPATH:BOOL=FALSE
  -DCMAKE_INSTALL_NAME_DIR:PATH=/path/to/install/dir