Alter name use of 'netgen' and 'netgen-mesher'
The name of the 'netgen' executable collides with an older UNIX pcb
trace routing application also named 'netgen'.  Fedora, for this
reason, renames this mesher package to 'netgen-mesher' and the
executable, likewise, to 'netgen-mesher'.  I propose the same change
here.

Ironically though, the current python module is already named
'netgen-mesher', which complicates loading of the module as 'netgen',
often resulting in it being unfindable.  For this reason, I similarly
reverse the usage, renaming the python module from 'netgen-mesher' to
'netgen'.